Change from GPL to MIT
When choosing the GPL license originally, the intent was to allow anyone to use the schema, and also to encourage open sharing of and updates or modifications made to it.

However, it is recognised that a strict reading of the terms of the GPL also technically requires users to share source code of their entire application. In the case of, for example, a traffic signal controller - this is clearly outside the scope of what was intended, and would have commercial and IP implications that might discourage users from using the schema.

We have decided to change to the permissive MIT license - so that the emphasis is on the schema being easy to adopt.

We still encourage feedback and suggested additions and modification to the schema.